feat: add autonomy proxy to update nodeConditions
