+++ pkg/yurthub/cachemanager/cache_manager_test.go
@@ -44,7 +44,6 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/tools/cache"
 
-	appsv1beta1 "github.com/openyurtio/openyurt/pkg/apis/apps/v1beta1"
 	"github.com/openyurtio/openyurt/pkg/projectinfo"
 	hubmeta "github.com/openyurtio/openyurt/pkg/yurthub/kubernetes/meta"
 	"github.com/openyurtio/openyurt/pkg/yurthub/kubernetes/serializer"
@@ -72,7 +71,7 @@ func TestCacheGetResponse(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	yurtCM := NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	yurtCM := NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	testcases := map[string]struct {
 		group        string
@@ -673,7 +672,7 @@ func TestCacheWatchResponse(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	yurtCM := NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	yurtCM := NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	testcases := map[string]struct {
 		group        string
@@ -1058,7 +1057,7 @@ func TestCacheListResponse(t *testing.T) {
 	if err != nil {
 		t.Errorf("failed to create RESTMapper manager, %v", err)
 	}
-	yurtCM := NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	yurtCM := NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	testcases := map[string]struct {
 		group        string
@@ -1651,7 +1650,7 @@ func TestQueryCacheForGet(t *testing.T) {
 	if err != nil {
 		t.Errorf("failed to create RESTMapper manager, %v", err)
 	}
-	yurtCM := NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	yurtCM := NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	testcases := map[string]struct {
 		keyBuildInfo storage.KeyBuildInfo
@@ -2378,7 +2377,7 @@ func TestQueryCacheForList(t *testing.T) {
 	if err != nil {
 		t.Errorf("failed to create RESTMapper manager, %v", err)
 	}
-	yurtCM := NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	yurtCM := NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	testcases := map[string]struct {
 		keyBuildInfo storage.KeyBuildInfo
@@ -3220,7 +3219,7 @@ func TestCanCacheFor(t *testing.T) {
 			defer close(stop)
 			client := fake.NewSimpleClientset()
 			informerFactory := informers.NewSharedInformerFactory(client, 0)
-			m := NewCacheManager(nil, s, nil, nil, informerFactory)
+			m := NewCacheManager(s, nil, nil, informerFactory)
 			informerFactory.Start(nil)
 			cache.WaitForCacheSync(stop, informerFactory.Core().V1().ConfigMaps().Informer().HasSynced)
 			if tt.preRequest != nil {
@@ -3344,67 +3343,6 @@ func TestIsListRequestWithNameFieldSelector(t *testing.T) {
 
 			if isMetadataNameFieldSelector != tc.Expect {
 				t.Errorf("failed at case %s, want: %v, got: %v", k, tc.Expect, isMetadataNameFieldSelector)
-			}
-		})
-	}
-}
-
-func TestSetNodeAutonomyCondition(t *testing.T) {
-	testcases := []struct {
-		name           string
-		node           *v1.Node
-		expectedStatus v1.ConditionStatus
-	}{
-		{
-			name: "case1",
-			node: &v1.Node{
-				Status: v1.NodeStatus{
-					Conditions: []v1.NodeCondition{
-						{
-							Type:   v1.NodeReady,
-							Status: v1.ConditionTrue,
-						},
-					},
-				},
-			},
-			expectedStatus: v1.ConditionTrue,
-		},
-		{
-			name: "case2",
-			node: &v1.Node{
-				Status: v1.NodeStatus{
-					Conditions: []v1.NodeCondition{
-						{
-							Type:   appsv1beta1.NodeAutonomy,
-							Status: v1.ConditionTrue,
-						},
-					},
-				},
-			},
-			expectedStatus: v1.ConditionTrue,
-		},
-		{
-			name: "case3",
-			node: &v1.Node{
-				Status: v1.NodeStatus{
-					Conditions: []v1.NodeCondition{
-						{
-							Type:   appsv1beta1.NodeAutonomy,
-							Status: v1.ConditionFalse,
-						},
-					},
-				},
-			},
-			expectedStatus: v1.ConditionTrue,
-		},
-	}
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			setNodeAutonomyCondition(tc.node, tc.expectedStatus, "", "")
-			for _, condition := range tc.node.Status.Conditions {
-				if condition.Type == appsv1beta1.NodeAutonomy && condition.Status != tc.expectedStatus {
-					t.Error("failed to set node autonomy status")
-				}
 			}
 		})
 	}
