+++ pkg/yurthub/proxy/local/local_test.go
@@ -66,7 +66,7 @@ func TestServeHTTPForWatch(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
@@ -158,7 +158,7 @@ func TestServeHTTPForWatchWithHealthyChange(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	cnt := 0
 	fn := func() bool {
@@ -243,7 +243,7 @@ func TestServeHTTPForWatchWithMinRequestTimeout(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
@@ -335,7 +335,7 @@ func TestServeHTTPForPost(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
@@ -415,7 +415,7 @@ func TestServeHTTPForDelete(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
@@ -482,7 +482,7 @@ func TestServeHTTPForGetReqCache(t *testing.T) {
 	}
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, nil, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, nil, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
@@ -635,7 +635,7 @@ func TestServeHTTPForListReqCache(t *testing.T) {
 	sWrapper := wrapper.NewStorageWrapper(dStorage)
 	serializerM := serializer.NewSerializerManager()
 	restRESTMapperMgr, _ := hubmeta.NewRESTMapperManager(rootDir)
-	cacheM := cachemanager.NewCacheManager(nil, sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
+	cacheM := cachemanager.NewCacheManager(sWrapper, serializerM, restRESTMapperMgr, fakeSharedInformerFactory)
 
 	fn := func() bool {
 		return false
